Atualizado Validacao dos dados

+++ src/controllers/AlbunsController.js
@@ -1,25 +1,35 @@
+const { validationResult } = require('express-validator');
 const Album = require('../models/Albuns');
 
 class AlbunsController {
   static adicionar = async (req, res) => {
     try {
-      const novoAlbum = {
-        titulo: req.body.titulo,
-        ano_lancamento: req.body.ano_lancamento,
-        link_youtube: req.body.link_youtube,
-        link_spotify: req.body.link_spotify,
-        link_deezer: req.body.link_deezer,
-        banda: req.body.banda,
-        gravadora: req.body.gravadora,
-      };
+      const validacao = validationResult(req);
 
-      await Album.create(novoAlbum);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const novoAlbum = {
+          titulo: req.body.titulo,
+          ano_lancamento: req.body.ano_lancamento,
+          link_youtube: req.body.link_youtube || null,
+          link_spotify: req.body.link_spotify || null,
+          link_deezer: req.body.link_deezer || null,
+          banda: req.body.banda,
+          gravadora: req.body.gravadora,
+        };
 
-      res.status(201).json({
-        mensagem: `Album ${novoAlbum.titulo} adicionado com sucesso!`,
-        dados: novoAlbum,
-        status: 201,
-      });
+        await Album.create(novoAlbum);
+
+        res.status(201).json({
+          mensagem: `Album ${novoAlbum.titulo} adicionado com sucesso!`,
+          dados: novoAlbum,
+          status: 201,
+        });
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -31,9 +41,18 @@ class AlbunsController {
 
   static exibirTodos = async (req, res) => {
     try {
-      const albuns = await Album.find();
+      const validacao = validationResult(req);
 
-      res.status(200).json(albuns);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const albuns = await Album.find();
+
+        res.status(200).json(albuns);
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -45,11 +64,27 @@ class AlbunsController {
 
   static exibirUm = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      const album = await Album.findById(id);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      res.status(200).json(album);
+        const album = await Album.findById(id);
+
+        if (album) {
+          res.status(200).json(album);
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe album com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -61,25 +96,43 @@ class AlbunsController {
 
   static atualizar = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      const atualizacaoAlbum = {
-        titulo: req.body.titulo,
-        ano_lancamento: req.body.ano_lancamento,
-        link_youtube: req.body.link_youtube,
-        link_spotify: req.body.link_spotify,
-        link_deezer: req.body.link_deezer,
-        banda: req.body.banda,
-        gravadora: req.body.gravadora,
-      };
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      await Album.findByIdAndUpdate(id, atualizacaoAlbum);
+        const album = await Album.findById(id);
 
-      res.status(200).json({
-        mensagem: `Album de id ${id} foi atualiado com sucesso!`,
-        dados: atualizacaoAlbum,
-        status: 200,
-      });
+        if (album) {
+          const atualizacaoAlbum = {
+            titulo: req.body.titulo,
+            ano_lancamento: req.body.ano_lancamento,
+            link_youtube: req.body.link_youtube,
+            link_spotify: req.body.link_spotify,
+            link_deezer: req.body.link_deezer,
+            banda: req.body.banda,
+            gravadora: req.body.gravadora,
+          };
+
+          await Album.findByIdAndUpdate(id, atualizacaoAlbum);
+
+          res.status(200).json({
+            mensagem: `Album de id ${id} foi atualiado com sucesso!`,
+            dados: atualizacaoAlbum,
+            status: 200,
+          });
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe album com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
@@ -91,14 +144,32 @@ class AlbunsController {
 
   static deletar = async (req, res) => {
     try {
-      const { id } = req.params;
+      const validacao = validationResult(req);
 
-      await Album.findByIdAndDelete(id);
+      if (!validacao.isEmpty()) {
+        res.status(400).send({
+          erro: validacao.array(),
+          status: 400,
+        });
+      } else {
+        const { id } = req.params;
 
-      res.status(200).json({
-        mensagem: `Album de id ${id} foi deletado com sucesso!`,
-        status: 200,
-      });
+        const album = await Album.findById(id);
+
+        if (album) {
+          await Album.findByIdAndDelete(id);
+
+          res.status(200).json({
+            mensagem: `Album de id ${id} foi deletado com sucesso!`,
+            status: 200,
+          });
+        } else {
+          res.status(400).json({
+            mensagem: `Não existe album com o id ${id}`,
+            status: 400,
+          });
+        }
+      }
     } catch (error) {
       res.status(500).json({
         mensagem: 'Ocorreu um erro interno no servidor!',
